PLEASE NOTE: Don't use this in production!
This is a project I use on my LAN ONLY, it does not interact with the internet at all!
It gets data from another application and the data never gets to this application if it's not safe to put in the database this way!
It will be updated in the future, but I didn't want to spend the time to do it correctly when I made it. Sorry!